Character navigation links for the Abyss Tracker profile page, for the characters' other fits, zKillboard page, and Eve Workbench page.
Ship navigation links for its Abyss Tracker info page, other fits with this ship, and again the two external links, zKill and Eve Workbench.

Firepower section - here I made the layout a bit different from Pyfa and Eve, so the total dps is highlighted and the 3 other stats are under each other in a list. (What is the measurement unit that goes with the Volley?)

Hitpoints section - here the layout is familiar enough - but I will probably change it to the progress bar style indicator from the pie graphs - this will also save some space

Repairs section: This is work in progress, but it won't be any different from Pyfa, apart from a few minor tweaks - I renamed the "Reinforced" repairs to "Burst" repairs and shield recharge will also be counted in there as well.